# docs/swagger.yaml
basePath: /api
definitions:
  models.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
        x-order: "2"
      username:
        example: admin
        type: string
        x-order: "1"
    type: object
  models.LoginResponse:
    properties:
      member:
        $ref: '#/definitions/models.Member'
      token:
        type: string
    type: object
  models.Member:
    properties:
      username:
        example: johndoe
        type: string
        x-order: "1"
    type: object
host: localhost:3000
info:
  contact: {}
  title: Member API with Fiber and GORM
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
      summary: Login to get JWT Token
      tags:
      - Auth
  /members:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Member'
            type: array
      security:
      - BearerAuth: []
      summary: Get all members
      tags:
      - Members
    post:
      consumes:
      - application/json
      description: สร้างสมาชิกใหม่โดยรับข้อมูลจาก JSON body
      parameters:
      - description: Member Data (username and password)
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Could not create member
          schema:
            additionalProperties: true
            type: object
      summary: Create a new member
      tags:
      - Members
  /members/{id}:
    delete:
      description: ลบสมาชิกออกจากระบบตาม ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a member
      tags:
      - Members
    get:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
      security:
      - BearerAuth: []
      summary: Get member by ID
      tags:
      - Members
    put:
      consumes:
      - application/json
      description: แก้ไขข้อมูลสมาชิก (เช่น เปลี่ยน Password) ตาม ID
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: New Member Data
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
      security:
      - BearerAuth: []
      summary: Update a member
      tags:
      - Members
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'ใส่ Token ในรูปแบบ: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"